+++ management_api_app/db/repositories/resources.py
@@ -1,21 +1,10 @@
-import uuid
-from typing import List
-
 from azure.cosmos import CosmosClient
-from pydantic import parse_obj_as, UUID4
+from pydantic import UUID4
 
 from core import config
-from resources import strings
 from db.errors import EntityDoesNotExist
-from models.domain.workspace import Workspace
 from db.repositories.base import BaseRepository
-from models.domain.resource import Deployment, Status, ResourceType
-from models.domain.resource_template import ResourceTemplate
-from models.schemas.workspace import WorkspaceInCreate
-from db.repositories.resource_templates import ResourceTemplateRepository
-from services.concatjsonschema import enrich_workspace_schema_defs
 from jsonschema import validate
-from services.authentication import extract_auth_information
 
 
 class ResourceRepository(BaseRepository):
@@ -23,9 +12,15 @@ class ResourceRepository(BaseRepository):
         super().__init__(client, config.STATE_STORE_RESOURCES_CONTAINER)
 
     @staticmethod
-    def _active_workspaces_query():
-        return 'SELECT * FROM c WHERE c.isDeleted = false'
-
-    @staticmethod
-    def _validate_workspace_parameters(workspace_create, workspace_template):
+    def _validate_resource_parameters(workspace_create, workspace_template):
         validate(instance=workspace_create["properties"], schema=workspace_template)
+
+    def get_resource_dict_by_id(self, workspace_id: UUID4) -> dict:
+        query = self._active_workspaces_query() + f' AND c.id="{workspace_id}"'
+        workspaces = self.query(query=query)
+        if not workspaces:
+            raise EntityDoesNotExist
+        return workspaces[0]
+
+    def update_resource_dict(self, resource_dict: dict):
+        self.container.upsert_item(body=resource_dict)
