+++ management_api_app/tests_ma/test_service_bus/test_deployment_status_update.py
@@ -74,10 +74,10 @@ async def test_receiving_good_message(app, sb_client, logging_mock, repo):
     sb_client().get_queue_receiver().complete_message = AsyncMock()
     expected_workspace = create_sample_workspace_object(test_sb_message["id"])
     expected_workspace.deployment = Deployment(status=Status.Deployed, message="")
-    repo().get_workspace_dict_by_workspace_id.return_value = expected_workspace.dict()
+    repo().get_resource_dict_by_id.return_value = expected_workspace.dict()
 
     await receive_message_and_update_deployment(app)
-    repo().get_workspace_dict_by_workspace_id.assert_called_once_with(uuid.UUID(test_sb_message["id"]))
+    repo().get_resource_dict_by_id.assert_called_once_with(uuid.UUID(test_sb_message["id"]))
     repo().update_resource_dict.assert_called_once_with(expected_workspace.dict())
     logging_mock.assert_not_called()
     sb_client().get_queue_receiver().complete_message.assert_called_once_with(service_bus_received_message_mock)
@@ -92,7 +92,7 @@ async def test_when_updating_non_existent_workspace_error_is_logged(app, sb_clie
 
     sb_client().get_queue_receiver().receive_messages = AsyncMock(return_value=[service_bus_received_message_mock])
     sb_client().get_queue_receiver().complete_message = AsyncMock()
-    repo().get_workspace_dict_by_workspace_id.side_effect = EntityDoesNotExist
+    repo().get_resource_dict_by_id.side_effect = EntityDoesNotExist
 
     await receive_message_and_update_deployment(app)
     expected_error_message = strings.DEPLOYMENT_STATUS_ID_NOT_FOUND.format(test_sb_message["id"])
@@ -109,7 +109,7 @@ async def test_when_updating_and_state_store_exception(app, sb_client, logging_m
 
     sb_client().get_queue_receiver().receive_messages = AsyncMock(return_value=[service_bus_received_message_mock])
     sb_client().get_queue_receiver().complete_message = AsyncMock()
-    repo().get_workspace_dict_by_workspace_id.side_effect = Exception
+    repo().get_resource_dict_by_id.side_effect = Exception
 
     await receive_message_and_update_deployment(app)
     logging_mock.assert_called_once_with(strings.STATE_STORE_ENDPOINT_NOT_RESPONDING + " ")
